ci: separate svg validation and optimization, refactor and add validate file path
